Track connection end
git-svn-id: file:///home/svn/DVA2006@5824 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/tasks/ACARSDataPurgeTask.java
@@ -1,4 +1,4 @@
-// Copyright 2005, 2006, 2007, 2008 Global Virtual Airlines Group. All Rights Reserved.
+// Copyright 2005, 2006, 2007, 2008, 2009 Global Virtual Airlines Group. All Rights Reserved.
 package org.deltava.tasks;
 
 import java.util.*;
@@ -16,7 +16,7 @@ import org.deltava.util.system.SystemData;
 /**
  * A Scheduled Task to purge old ACARS log data.
  * @author Luke
- * @version 2.3
+ * @version 2.4
  * @since 1.0
  */
 
@@ -34,46 +34,49 @@ public class ACARSDataPurgeTask extends Task {
 	 */
 	protected void execute(TaskContext ctx) {
 		log.info("Executing");
-		
+
 		// Get active flights
 		GetACARSPool acdao = new GetACARSPool();
 		Collection<Integer> activeIDs = acdao.getFlightIDs();
-		
+
 		// Determine the purge intervals
 		int flightPurge = SystemData.getInt("log.purge.flights", 48);
-		int conPurge = SystemData.getInt("log.purge.cons", 60);
+		int conPurge = SystemData.getInt("log.purge.cons", 48);
 		int statsPurge = SystemData.getInt("log.purge.acars_stats", 96);
 		try {
 			Connection con = ctx.getConnection();
-			
+
 			// Ensure all archived data is in the right place
 			SetACARSLog wdao = new SetACARSLog(con);
 			wdao.synchronizeArchive();
-			
+
 			// Remove old flights and position reports without a flight report
 			Collection<Integer> purgedIDs = wdao.purgeFlights(flightPurge, activeIDs);
 			log.warn("Purged " + purgedIDs.size() + " flight entries - " + purgedIDs);
-			
+
 			// Purge old stats
 			log.warn("Purged " + wdao.purgeLogs(statsPurge) + " command statistics entries");
-			
+
 			// Get connections
 			GetACARSLog dao = new GetACARSLog(con);
 			Collection<ConnectionEntry> cons = dao.getUnusedConnections(conPurge);
-			
+
 			// Purge the connections
 			int purgeCount = 0;
-			for (Iterator<ConnectionEntry> i = cons.iterator(); i.hasNext(); ) {
+			for (Iterator<ConnectionEntry> i = cons.iterator(); i.hasNext();) {
 				ConnectionEntry ce = i.next();
-				try {
-					wdao.deleteConnection(ce.getID());
-					purgeCount++;
-					log.info("Purged Connection " + StringUtils.formatHex(ce.getID()));
-				} catch (DAOException de) {
-					log.error("Error purging Connection " + StringUtils.formatHex(ce.getID()) + " - " + de.getMessage());
+				if (!ce.getDispatch()) {
+					try {
+						wdao.deleteConnection(ce.getID());
+						purgeCount++;
+						log.info("Purged Connection " + StringUtils.formatHex(ce.getID()));
+					} catch (DAOException de) {
+						log.error("Error purging Connection " + StringUtils.formatHex(ce.getID()) + " - "
+								+ de.getMessage());
+					}
 				}
 			}
-			
+
 			// Log purge count
 			log.info("Purged " + purgeCount + " connection entries");
 		} catch (DAOException de) {
